Move SQLiteCommandED etc. to separate source file
